feat: add comprehensive support of cityjson, cityjsonseq, and flatcitybuf (#4)
* feat: add remote file reader, FlatCityBuf support, and COPY TO

Task 1 - Remote File Reader:
- Add ReadFileContent() using DuckDB FileSystem API for HTTP/S3/GCS support
- Add content-based constructors to LocalCityJSONReader and LocalCityJSONSeqReader
- Add context-aware OpenAnyCityJSONFile() factory overload
- Update bind functions and metadata functions to use DuckDB FileSystem

Task 2 - FlatCityBuf Reader (guarded by CITYJSON_HAS_FCB):
- Add FlatCityBufReader class implementing CityJSONReader interface
- Add read_flatcitybuf and flatcitybuf_metadata table functions
- Add CMake FetchContent for flatcitybuf library (opt-in via CITYJSON_ENABLE_FCB)
- Add .fcb extension detection in reader factory

Task 3 - COPY TO Statement:
- Register CopyFunction for 'cityjson' and 'cityjsonseq' formats
- Implement sink/combine/finalize pipeline for accumulating CityObjects
- Add CityJSONWriter with vertex quantisation for both formats
- Support metadata_query option and explicit version/crs/transform options
- Column role detection maps input columns to CityJSON structure


* fix: bug of copy statement

* fix: populate all metadata fields in COPY statement

- Fix dangling reference in MetadataScan by using SetValue instead of
  Reference, which caused use-after-free when COPY consumed output after
  the temporary DataChunk was destroyed
- Parse geographicalExtent as array of 6 numbers (not string) matching
  the CityJSON spec
- Populate geographical_extent column from parsed metadata instead of
  hardcoding NULL
- Use metadata.referenceSystem URI as fallback when top-level CRS is
  not present (CityJSON 2.0 style)
- Reorder GeographicalExtent struct before Metadata in header to fix
  forward declaration dependency


* feat: copy statement of cityjson writer

* chore: change duckdb version to stable

* feat: flatcitybuf IO support

* test: e2e and clean up docs

* chore: format

* format

* ci: default flatcitybuf flag

* ci: fix

* ci: fix

* ci: replace FCB test with no-op when FCB is disabled

On platforms where FlatCityBuf is auto-disabled (old glibc, musl),
the FCB test file would still run and fail because the flatcitybuf
functions aren't registered. This adds an else() block that overwrites
the test file with a harmless SELECT 1 at CMake configure time.

* chore: duckdb v1.5.0

* ci: fix FCB builds on macOS x86_64, Windows, and Wasm

Three issues fixed:

1. macOS osx_amd64: CI runs on arm64 host but targets x86_64.
   CMAKE_SYSTEM_PROCESSOR reports the host arch, so the wrong
   (aarch64) binary was downloaded. Now checks CMAKE_OSX_ARCHITECTURES
   first for correct cross-compilation detection.

2. Windows: cmake -E tar xzf (gzip flag) may not extract .zip
   archives correctly. Changed to xf (auto-detect format). Also
   added post-extraction verification that the library file exists.

3. Wasm/Emscripten: FATAL_ERROR on unsupported platform now
   gracefully disables FCB instead of failing the entire build.

All platform/arch detection errors are now non-fatal — they
disable FCB and fall through to the no-op test file generation.

* ci: disable FCB on macOS x86_64 (upstream binary is arm64)

The fcb_cpp-macos-x86_64.tar.gz release asset from flatcitybuf
v0.7.6 actually contains an arm64 binary, causing linker failures
when cross-compiling for x86_64 on Apple Silicon CI runners.

Disable FCB for this target until the upstream release is fixed.

---------

+++ src/cityjson/temporal_parser.cpp
@@ -105,7 +105,7 @@ int64_t ParseTimestampString(const std::string &timestamp_str) {
 
 		// Convert to DuckDB timestamp (microseconds since epoch)
 		date_t date = Date::FromDate(year, month, day);
-		dtime_t time = Time::FromTime(hour, minute, second, microseconds);
+		dtime_t time = Time::FromTime(hour, minute, second, static_cast<int32_t>(microseconds));
 
 		// Combine date and time
 		timestamp_t timestamp = Timestamp::FromDatetime(date, time);
@@ -114,7 +114,7 @@ int64_t ParseTimestampString(const std::string &timestamp_str) {
 		if (match[8].matched && match[9].matched) {
 			int tz_hour = std::stoi(match[8].str());
 			int tz_minute = std::stoi(match[9].str());
-			int64_t tz_offset_micros = (tz_hour * 60 + tz_minute) * 60 * 1000000LL;
+			int64_t tz_offset_micros = static_cast<int64_t>(tz_hour * 60 + tz_minute) * 60LL * 1000000LL;
 			timestamp.value -= tz_offset_micros;
 		}
 
@@ -168,7 +168,7 @@ int64_t ParseTimeString(const std::string &time_str) {
 		}
 
 		// Convert to DuckDB time (microseconds since midnight)
-		dtime_t time = Time::FromTime(hour, minute, second, microseconds);
+		dtime_t time = Time::FromTime(hour, minute, second, static_cast<int32_t>(microseconds));
 		return time.micros;
 
 	} catch (const std::exception &e) {
